ensure that defintion position of type parameters is not lost when adding bounds
Summary:
This solves a problem where calling `Env.add_upper_bound` or `Env.add_lower_bound` would cause the definition position of the type parameter to be lost and replaced by `Pos.none`.

This was in part due to `Type_parameter_env.add` taking the definition position as an optional parameter, even though it should always be set. I've therefore made this a mandatory parameter.

Reviewed By: andrewjkennedy

Differential Revision: D23341824

fbshipit-source-id: 57debc9a3552d7dc192cb1ccc8ba0de384112f14