Update VariableSerializer to correctly handle legacy bit with Eval.HackArrDVArr on
Summary:
As title.

serialize_impl() proves is a bit of bug-factory due to hand-coding serialization for empty arrays.
I was tempted to remove fast path for keepDVArrays completely, but it is likely to be perf regression - it's called from memcache client, which is hot and encounters lot of empty array values.

Reviewed By: arnabde03

Differential Revision: D23309027

fbshipit-source-id: e13379c4721b5a5d9ae219498287b03ce9ca4ea6