Don't pipe files to typecheck into next fanout calculation request
Summary:
Every time we advance the cursor, we take all the changed files and calculate their fanout. We batch up all the file changes in this way for a later calculate-errors operation.

Before this commit, we were including both the changed files and the fanout of the previous cursor advancement. That's completely unnecessary and blows up the fanout size. Only the `calculate_errors` method should be calling `get_files_to_typecheck`, because it's the only method doing typechecking.

Differential Revision: D23261938

fbshipit-source-id: a54452ca0a5ae92c4083b0ee2bf85e6cf6b33142