Settle on only having one `hh_fanout` implementation
Summary: In principle, I wanted to have it be possible to select a new backend for incremental state at runtime. I don't think that's likely to happen; instead, we'll just have one implementation for the state. This commit just collapses the abstractions down into one file and makes sure it compiles/passes tests. (We no longer need classes/objects either, but I don't address that in this diff.)

Reviewed By: 2BitSalute

Differential Revision: D23086219

fbshipit-source-id: b926dbb1c70c24376d3f8643c4ffb599477822bb