Don't query Watchman for saved-state when given a previous cursor
Summary:
We currently indiscriminately query Watchman for saved-state information on every request. We don't need to do this if we've been given the previous cursor — all the saved-state information is already stored inside that cursor.

This commit is mostly refactoring. The most interesting thing is that we now lazily open the connection to the dep-table SQLite saved-state when we try to advance the cursor, not when we query Watchman for saved-state information. (This is now necessary because we no longer query Watchman in every code path. We don't need to do this for the naming-table saved-state since it's not a global object that needs initialization — we directly pass it in where appropriate.)

Reviewed By: 2BitSalute

Differential Revision: D23277774

fbshipit-source-id: 8f00a13ccb6e0a1536326415b3f51ddc2eecf75e