Don't store hashes in client IDs
Summary:
There's a bit of misguided code where we modify the client ID that was passed to us and sign it with a hash. We don't do this anywhere else, and we actually need to look up data based on the unsigned client ID. Just delete that bit of code to fix things up.

Note that cursor IDs *are* signed with a hash of the associated client configuration, which is where the confusion came from.

Reviewed By: 2BitSalute

Differential Revision: D23277775

fbshipit-source-id: 07d81cf88e40425cdc463aa1ff4efdefdef6c780