Add hh\prefetch_units
Summary:
Add a builtin that allows one to manually request the prefetching of
units. A request can either be advisory or non-advisory. An advisory
request is just a hint and can be ignored by the runtime. Even if we
prefetch the units, it may be done asynchronously at a later
point. Non-advisory requests, on the other hand, will block until the
prefetch attempts finish. Non-advisory requests may be slower than
advisory requests, depending on the configuration (if unit prefetching
isn't enabled, we have to do it inline).

Reviewed By: oulgen

Differential Revision: D23288499

fbshipit-source-id: 45ff1380016090bd4c6d92bcf92f622fe4b3ad38