Simplify bccache and remove HHAS caching mode
Summary:
Remove the ability of bccache to cache HHAS. From now on, it will
cache UnitEmitters only, as that's proven always superior. Remove a
lot of unused/unneeded cruft from bccache, including several unused
builtins.

Add in support for limiting the total fetch time, which had a
configuration but wasn't hooked up to anything.

One of the side-effects of not caching HHAS is that its not possible
to implement the fetching builtins for bccache. These are left empty
for the time being, but will be reimplemented later on the diff stack
using a more generic prefetching mechanism.

Reviewed By: oulgen

Differential Revision: D23245675

fbshipit-source-id: 5a4a76f6aaf5895660340b481d7954623038002d