Improve HackC worker management in extern-compiler.cpp
Summary:
Instead of having an AtomicVector of HackC worker threads, and
iterating over the array to find a free worker, just use a std::vector
instead. Use the vector like a stack, popping off the first available
worker, and pushing it back on when done. This simplifies the
book-keeping (we don't need to remember indices) and makes it more
likely to re-use a worker just used (which is better for caching). The
locking can also be simplified. We don't need any fancy atomics, just
a conventional lock and condition variable for when the vector is
empty.

Reviewed By: mofarrell

Differential Revision: D23261620

fbshipit-source-id: d3a0640498a368dcdd54689e4d1482868cdfa901