Move assembler call out from under HackC worker lock
Summary:
Right now we call the assembler immediately after getting
HHAS from HackC. This means we invoke the assembler (which isn't
particularily fast) while holding the worker lock. This increases lock
contention unnecessarily since we don't need the worker lock for
this. Refactor the pipeline a bit to invoke the assembler after
releasing the lock.

Reviewed By: paulbiss

Differential Revision: D23250122

fbshipit-source-id: 7497cadf540a9b5b902dace0b7c9941da72c77fa