Move bccache hook out of HackC logic
Summary:
Right now we invoke the bccache hook deep within the logic to call out
to HackC. This is unfortunate placement since there's no intrinsic
reason why the hook should be so connected to HackC. Also, it means we
invoke the hook while holding the lock for the HackC workers, which is
inefficient (increases contention).

Instead move up the hook. Now the caching hook is its own type of
UnitCompiler. When we ask for a UnitCompiler, it will either give us a
HackC unit compiler, or a caching unit compiler (depending on whether
the hook is enabled). The caching unit compiler will fallback to the
HackC compiler if the hook fails.

The advantage is we only hold the HackC workers locks when we're
actually using HackC, not when using the hook. It also simplifies the
logic a bit. This required a bit of work to handle ICE UnitEmitters
properly. If we generate a UnitEmitter with an ICE inside the caching
hook, we don't want to cache it, so we tell the HackC UnitCompiler to
just return null in that case.

Also rename the hook to make it clear the hook is for caching, not
just a general compiler hook.

Reviewed By: oulgen

Differential Revision: D23249614

fbshipit-source-id: dee2a79091f239377115b275067577cb308aa657