Implement prefetching of units
Summary:
Implement a unit prefetcher. If configured, this will attempt to
asynchronously load units (in separate threads) that seem likely to be
needed in the future. The benefit being, of course, that when the unit
is actually needed, it will already be available and the request does
not need to block loading it.

The mechanism works as follows: We maintain a pool of worker threads
(using folly executors). Eval.UnitPrefetcherMaxThreads and
Eval.UnitPrefetcherMinThreads controls the maximum and minimum size of
the thread pool. Threads are created as long as there's available
work, up to the max. Threads idle for longer than
Eval.UnitPrefetcherIdleThreadTimeoutSecs are reaped (down to the
minimum). Setting Eval.UnitPrefetcherMaxThreads to 0 disables unit
prefetching. Unit prefetching is currently only available for
non-RepoAuthoritative mode.

The heuristic we use to determine which units to prefetch is
relatively simple. If we load a unit (that is, from an actual
request), we then prefetch any unit statically referenced by the unit
being loaded. This is obtained by the symbol ref info from HackC. This
won't catch dynamic references, but it catches most.

We now always request symbol refs from HackC. The symbol refs we parse
from any file are stored in the UnitEmitter and the eventual Unit. The
Unit also includes an atomic flag if the symbol refs have been
"claimed". When loading an unit (that is, from an actual request), we
check if the symbol refs have been claimed. If not, we claim them
(atomically), then pass the symbol refs in the autoloader to map them
to paths. We then enqueue the paths into the worker pool queue to
prefetch them asynchronously.

As an additional optimization, we use a "prefetch version" counter to
quickly avoid enqueueing prefetch requests for units which we know
haven't changed. Many prefetch requests using the above mechanism are
duplicates, so this saves a lot of overhead.

The actual work of prefetching the unit is pretty similar to the
normal unit load path. We avoid dealing with problematic corner cases
(streams for example). It's always safe to drop a prefetch request,
since the prefetching is just an optimization. We also attempt to grab
the unit lock without blocking. If the lock is already required,
another thread is already loading the same unit, so there's no point
in blocking. If we fail to acquire the lock, we just give up.

Add some intrinsics to make testing the prefetcher easier, and provide
new unit tests to test it.

Reviewed By: mofarrell

Differential Revision: D23288495

fbshipit-source-id: 137f209a9930c544b5522555c3268937a0af7075