proper scoping of type parameters in type constraints during naming
Summary:
This is a step towards allowing constraints on higher-kinded type parameters and nested type parameters (i.e., type parameters of type parameters).

In `Naming`, existing `Happly (x, args)` hints are turned into `Habstr (x, args)` hints, if `x` is a type parameter. `Habstr` is the hint equivalent of `Typing_defs_core.Tgeneric`.

Therefore, `Naming` must be aware of the proper scoping of (nested) type variables, in order to do the transformation described above on the constraints found on such type parameters. This diff implements this.

The scoping is described in a lengthy comment on the function `Naming.type_param`.

Reviewed By: andrewjkennedy

Differential Revision: D23244429

fbshipit-source-id: 7158a913999c01c8b0de19fae6b64a4ebd979823