Split up LoggingArray and LoggingProfile
Summary:
As discussed in D23302834 (https://github.com/facebook/hhvm/commit/075d55bf392f62b8a2d5787e25c6c5eab9310c54), we should split up these two pieces. It'll be especially important as we address all the issues in T74326631. LoggingArray already has enough complexity dealing with array invariants and refcounting; LoggingProfile should deal with input formatting and thread safety.

logging-profile.cpp is responsible for tracking event updates, storing the global profile map, and exporting the map. It exposes a constrained API for events, which will be important in the next diff, which packs event keys.

I intended for this diff to be pure code motion, but I found a few logical issues that were  easier to fix than to migrate as-is.
1. In convert, we should count keyset -> keyset as another "layout unchanged - keep the same profile" case.
2. There's a thread-safety issue with creating a LoggingProfile: one thread does the insert, and if it succeeds, it sets the value. If another thread reads the nullptr LoggingProfile in between, we'll crash. Fix that by using the insert(pair) form instead - see getLoggingProfile in logging-profile.cpp.
3. There's a thread-safety issue with creating a new staticArray: one thread does the insert, and if it succeeds, it sets the value. Fix it by making staticArray an atomic field and then compare-exchanging it - see profileForSrcKey in logging-array.cpp.

There is still a significant thread-safety bug here; we can't iterate over the hash map in order to export it while things are still writing to it, and it's essentially impossible to block writes. (LoggingArrays can get stored in APC, and even if we tried to block writes by setting an atomic flag, someone could be past the flag check but before the write.) I left notes on how to fix that as a follow-up.

Reviewed By: ricklavoie

Differential Revision: D23366735

fbshipit-source-id: cbd03711d2ba83e1452ba150c0508e6ce627a88c