Properly unwind stack in fcallHelper()
Summary:
Fixups and catch blocks of Call instructions expect the implementation to
unwind the stack so that it no longer contains the space reserved for ActRec
and inout arguments. This was not the case if fcallHelper() encounters a stack
overflow. This was "fixed" by writing NullUninits over the space and expecting
the unwinder to be able to deal with it.

This could potentially cause an issue with jitted unwinder implementation, such
as the one attempted by oulgen.

Let's instead manually unwind the stack to the point expected by the Call
opcode. Removes the necessity to eagerly write NullUninits.

Also, abstract this logic and all logic related to dealing with VM regs to a
separate helper, so that it can be reused by the next diff.

While there, use a proper name for numArgs to indicate whether it includes the
unpack argument, and fix a bug of passing a number that includes unpack to the
`func->numInOutParamsForArgs()` helper. This was benign, as if we have unpack
arguments in fcallHelper(), they are guaranteed to be at the correct position
and thus do not affect the inout count. It won't be true with the helper
introduced by the next diff though.

Reviewed By: ricklavoie, oulgen

Differential Revision: D23338745

fbshipit-source-id: 090db648f8de165e3445e2707f867243ee63e321