Implement generic unpack support in Call instruction
Summary:
Add a funcPrologueRedispatchUnpack stub that repacks arguments and transfers
control to the prologue. Use it by the Call opcode to implement generic unpack
support. Use Call opcode instead of CallUnpack.

It is possible to implement this more efficiently without having to sync the
VM registers, but the most common use cases are already handled by repacking
arguments in irgen. It is also doing strictly less work than CallUnpack and the
implementation is simpler.

Reviewed By: ricklavoie

Differential Revision: D23338796

fbshipit-source-id: 20ec43856b65838b1f9b6cf2b29c31400f67e6ee