Stop logging for arrays after RTA is scheduled
Summary:
When we output the data in our LoggingProfiles, we aim to have full profiles (i.e. all requests during which LoggingArrays were emitted have completed). We output our profiles when retranslateAll is reached. Therefore, we stop emitting new LoggingProfiles right before retranslateAll is placed on the treadmill. However, some LoggingArrays may stick around after this is done (e.g. in APC). We do not want to record events for these, as we use non-thread-safe iterators to analyze the logging data. Specifically, we need to ensure that when the export begins, `s_profileMap` and the profile-specific event maps will never be modified again.

We only allow events to be recorded if s_exportStarted is false. We protect s_exportStarted with a SharedMutex which we acquire for reads when we record events and acquire for writes when we set the flag. When we are about to record an event, we acquire the mutex for reads, check s_exportStarted, and record only if s_exportStarted is false. The read holder is released only once the log operation is complete.

When we begin outputting profile data, we acquire the mutex for writes, set s_exportStarted, and release the mutex. This ensures that when the mutex is released, no further updates will occur on the profile maps. We are free to perform non-thread-safe operations without needing to swap the map.

We also refactor so that no incomplete LoggingProfiles (i.e. those with uninitialized staticArrays) are ever added to the map.

In summary, we do the following:
1. Set g_emitLoggingArrays to false so that no further logging arrays will be emitted. s_profileMap will no longer be modified, except for logging arrays that are currently being created. These modification operations must complete before their request completes.
2. Place the export operation on the treadmill so that all request-bound logging arrays live out their full life.
3. The export operation begins, acquiring the write lock on s_exportStarted and setting it to true. No mutations are in progress because the write lock has been acquired.
4. All modifications to profile-specific event maps are guarded on s_exportStart being false (with a read lock on the mutex for the duration of the modification operation) and no longer succeed.

Differential Revision: D23347915

fbshipit-source-id: b9c45381786e5d6fca9164b6807c1ad1f497e0af