Remove Typing_defs.class_const.cc_expr
Summary:
I think of "declarations" as "summaries of AST", and for the most part it works like that: we take complicated  AST construct and boil it down to a type, or boolean flag, or a simpler variant.

`cc_expr` and `scc_expr` are two major eyesores in this idealized picture: the only places where we embed part of AST directly into declarations. This forces direct decl constructors to re-encode entire grammar of constant expressions into one more place, which we could do, but it will be super annoying to get all the edge cases right, and annoying to keep maintaining it in the future.

Fortunately, there are literally two consumers of this field, that I remove in dependent diffs.

Note that we will still have to look at expressions in places where we "infer" missing typehints and similar, but this is much less critical in an increasingly strict codebase.

Reviewed By: dlreeves

Differential Revision: D23334849

fbshipit-source-id: 41670ddaa6bdaf4368f12fe901d8cce5ebaae07c