Strip valid markdown delimiters in error messages sent from client LSP
Summary:
This change removes markdown delimiters in error messages sent from client LSP to VSCode, because VSCode is not currently rendering markdown text in the hover hints (i.e. it leaves them as they are).

The change renders the error message, which has the effect of producing a string with escape characters, which when displayed by VSCode appears to just be printing the visible characters as desired.

Reviewed By: Wilfred

Differential Revision: D23375945

fbshipit-source-id: 3755b074192dbeb990bd3e772e3922b1785b6ba5