Add tilde as a markdown delimiter
Summary:
Adds the double tilde (`~~`) as a markdown delimiter for error messages. Since strikethrough isn't supported in VSCode terminals, this delimiter instead highlights the delimited text in red.

It sets color of highlighted text to match  the marker color, unless the marker color is yellow (which is only the case for Lints, and only occurs in the main message of an error. Since Lints typically only have the main message (no subsequent reason messages) and because yellow is not part of the color wheel for reason messages regardless, it is unambiguous to use Red in that case.

Reviewed By: Wilfred

Differential Revision: D23381624

fbshipit-source-id: 7c276d6dee92e656e6c435913dd325e447f4ff67