Weight logging array sample counts by effective sampling ratio
Summary:
The raw event sample counts are currently biased based on how many times we emit a logging array. This is especially pronounced for arrays that are constructed a small number of times, or arrays that are the product of certain kinds of conversions.

The emitted event counts are now estimated *total* event counts had a logging array been emitted each time. We also weight convert profiles correctly given that they are only emitted when a logging array reaches them.

Differential Revision: D23357871

fbshipit-source-id: d6561b359cb9c217914e18bc39efa3201a2a419d