Remove savedGenerics from doFCall()
Summary:
Re-align doFCall() with the functionality of the prologue and remove the
optimization for cold interpreter code path that tries to avoid an unnecessary
check for errors in reified generics. This optimization makes interpretation of
all other FCalls tiny bit slower since they have to check for this condition.
It also makes it more difficult to improve the calling convention.

Just pass generics via stack as both doFCall() and prologue expect it.

Reviewed By: paulbiss

Differential Revision: D23415915

fbshipit-source-id: 5932405787dfd0fdc6ab03ce2de1d9cf3c27cf87