Repack arguments as needed before calling doFCall() in interpreter
Summary:
When doFCall() is called from the stub via fcallHelper(), it already receives
properly packed arguments. Move the argument repacking logic to its only other
caller used by the interpreter.

After this diff the responsibilities of doFCall() match those of function
prologues, which makes it easier to rearrange the logic of prologues and
doFCall() in sync.

Reviewed By: oulgen

Differential Revision: D23420377

fbshipit-source-id: 0ec5245aa1741e75701cb6adf688cb87e585f8b4