Audit and fix some legacy mark issues
Summary:
We'd like to start applying a "legacy mark" to some dvarrays in our codebase to suppress Hack Arr Compat logs. This mark should have no visible effect other than suppressing logs.

One thread of evidence for the safety of marking is a complete audit of places where it's used. The mark (as read by isLegacyArray()) is used in very little logic. I've collected a list of all of the behaviors it changes:
1. It's propagated by the PackedArray, MixedArray, and LoggingArray constructors.
1. It's propagated through GetScalarArray (so a given static array may have two entries - one marked, and one unmarked).
1. It's propagated through APC, but only for vecs and dicts. (We should propagate it for dvarrays, too.)
1. Legacy arrays don't get provenance tags.
1. is_array_marked_legacy returns true for legacy arrays.
1. Serialization logging (e.g. is_array, json_encode, serialize) is suppressed for legacy arrays.
1. Shape and tuple type-test logging is suppressed for legacy arrays.
1. VariableSerializer and serialize behave differently for legacy arrays.
1. Legacy arrays do not compare equal to non-legacy arrays.

Behaviors 1, 2, and 3 don't really count as "reads" of the legacy bit - all we do is read the bit on one array and write it on a copy. I rewrote array copy constructors to use auxBits instead, to make it clear that there's no actual logic based on the mark happening here.

Everything else on the list makes sense, except behavior 9, which I believe is just wrong. It would be close to correct if the behavior change was restricted to post-HADVAs, but I believe we have an alternate solution planned for comparison: we raise a Hack Arr Compat notice for PHP-Hack array comparison, and we can replace all sites raising this notice to call an observing helper, then ban the behavior. See here for logs: https://fburl.com/scuba/php_errors/k6pzvwza

In this diff, we wipe out behavior 9, and we also wipe out a related behavior where we read the provenance tag during comparison, too. I don't believe the tag comparison change is safe.

Reviewed By: arnabde03

Differential Revision: D23420013

fbshipit-source-id: 5292430f732ca962c218e8ca0a548d01e97dca6d