Kill AuxUnion u_deepInit
Summary: The only thing using u_deepInit in AuxUnion is code for converting Class::PropInitVec::Entry to/from a TypedValueAux. That in turn is only used in Class::sortOwnPropsInitVec to make a temporary copy of a range of Class::PropInitVec::Entries. For this use case, we don't need to try to shoehorn a TypedValue and a bool into a TypedValueAux, we can just store the TypedValue and bool separately. Do so, remove u_deepInit from AuxUnion, and remove the machinery for converting Class::PropInitVec::Entry to/from a TypedValueAux.

Reviewed By: oulgen

Differential Revision: D23424575

fbshipit-source-id: 53449b26c050c58a69ed1a87230cc478f234312f