Simplify policied property and generic treatment
Summary:
We simplify `Tclass` by making all properties be subject to the lump policy or have explicit purposes.

This means we no longer require flows in generics as properties using the generic type either have a purpose or completely subject to the lump policy.

It also means we don't have to do subtyping for policied properties as we already invariantly flow the lump policies and purposes generate trivial constraints of the form `A < A`.

An unfortunate consequence is that some errors are duplicated because they have different position sets. This should be easy to resolve.

Reviewed By: mpu

Differential Revision: D23399548

fbshipit-source-id: 9837070eb2f9177a10f80f930e2cce90fe35c220