Better error reporting for missing types in type constant access
Summary: In a type constant selection expression such as `C::T::T2::T3`, if a class is missing, or unparseable, during checking of the selection, we emit an "unbound name" error. (For example, if `C` does not exist, or if `C::T` is bound to a class that itself does not exist, etc.)  But the position in the error is of the eventual unbound class, not of the type selection itself. Fix this by using the position from the selection itself as the primary error, with optional secondary error pointing to the class that is actually missing.

Reviewed By: vsiles

Differential Revision: D23446482

fbshipit-source-id: 83b08ac172b87edda01a04f95a61ac7863a60d6d