Remove fixme_codes field from declarations
Summary:
This is unused, and diff that was going to use it (D15918881) stalled a year ago. While the problem this was going to solve (T46105464) probably still exists, in the meantime it got mitigated by lint rule (D18544985) that seems pretty successful - this field is empty in entire www aside from the test file for that lint rule.

This single file is also one of the last remaining mismatches for direct decl smart constructors (T71787903), and reimplementing fixme generation logic there doesn't look worth it if there are no concrete plans to use them.

Reviewed By: CatherineGasnier

Differential Revision: D23439278

fbshipit-source-id: 7b641d43a56dc17d49e6a89a71d2bcf617f6654e