Parsing of temporary capability syntax
Summary:
Adds the following syntax to be placed between a function's arguments and the return
```
@{ ty1 [+ ty2] }
```
Here, `ty1` represents the capability type of a function, and `ty2` is the optional unsafe capability that is intersected with the function's capability during function calls.
```
function f()@{ int }: void {
  g(); // ok
}
function g()@{ arraykey }: void {
  f(); // error, arraykey </: int
}
function g_unsafe()@{ arraykey + int }: void {
  f(); // ok, (arraykey & int) <: int
}
```
This syntax allows for granular prototyping and testing of coeffects and their applications. This feature is gated behind the experimental features flag `capabilities_provisional`. To use it, add
```
<<file: __EnableUnstableFeatures('capabilities_provisional')>>
```
to a Hack file.

Adding a node to the CST results in a massive number of .exp changes. The relevant changes in this diff are in `declaration_parser.rs`, `rust_parser_errors.rs`, and `schema_definition.ml`.

Reviewed By: losvald

Differential Revision: D22687006

fbshipit-source-id: eaf31862275c8610e50497cf22de8177396b1b48