Don't allow copying VarNR
Summary: I got a lint on D23424576 (https://github.com/facebook/hhvm/commit/d8461aae6edd3e2afc6a65c33435a463352edf59) that VarNR had a copy constructor but not a copy assignment operator and we should add the missing operator. I didn't think we should be copying VarNR: it is intended as a way to transform among some of the C++ types we use in HHVM, not as a way of storing things (which might require being able to copy and/or move). Indeed, we can `= delete` the copy constructor (and the copy assignment operator for good measure) and nothing breaks. Do so.

Reviewed By: oulgen

Differential Revision: D23434399

fbshipit-source-id: 033470f1ba47479fdf1205e916aa26f6fce37751