Push type/function lowercasing into UserAutoloadMap
Summary:
PHP treats types and functions in a case-insensitive manner, such that `new eNtPoCkEtMoNsTeR()` will create a new instance of `class EntPocketMonster {}` with no complaints or problems whatsoever. This continues to be true for the Hack runtime as well, though incorrectly-casing a type or function will throw a typechecker error, and we may eventually make Hack into a case-sensitive language.

The Hack autoloader has historically solved this problem by only storing lowercased copies of each type and function name in the AutoloadMap. This meant that our AutoloadMap doesn't actually know the correct capitalization of any type or function, though.

Technology has progressed, and now we have three AutoloadMaps:

* UserAutoloadMap (used in `/var/www` and open-source code)
* RepoAutoloadMap (used in repo-authoritative mode, in production)
* WatchmanAutoloadMap (used on some FB sandboxes)

`RepoAutoloadMap` and `WatchmanAutoloadMap` use case-insensitive comparators and are able to store correctly-capitalized strings. `UserAutoloadMap` still stores a standard userspace `Array` and looks each string up in a case-sensitive manner. But since this lowercasing logic is only needed for `UserAutoloadMap`, we can just push the lowercasing logic into `UserAutoloadMap` and avoid the effort in our other maps.

Reviewed By: jano

Differential Revision: D23438816

fbshipit-source-id: e85226efae6b8eb10f443e8d7ed4ec27b36ae1e3