refactor handling of governed functions
Summary: I've refactored the handling of governed functions in order to more nicely handle things that are governed by something other than an existential policy. The Governed attribute optionally takes a string argument. If you pass in a string, then the function is governed by a concrete purpose (ie, all inputs and outputs have that purpose). If you do not pass an argument, then it is governed by a universally quantified policy.

Reviewed By: mpu

Differential Revision: D23379902

fbshipit-source-id: 0d67fdf9f7ab63cbdb65622a20cfe48a1674b9b8